client/requirements.txt
grpcio==1.71.0
Pillow==11.2.1
protobuf==6.30.2
streamlit==1.45.0
